requirements.txt
bilibili_api_python==16.2.0
python-dotenv==1.0.1
pyinstaller==6.7.0
